agents/orchestrator/requirements.txt
a2a-sdk[http-server]>=0.2.0
google-generativeai>=0.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
uvicorn>=0.32.0
fastapi>=0.115.0
